+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## [3.7.0](https://gitlab.konghack.com/GCWorld/Routing/compare/3.6.2.1...3.7.0)
+ - @GameCharmer Added support for "autoWrapper" per new interface
+ - @GameCharmer Optimized code
+
+
 ## [3.6.2.1](https://gitlab.konghack.com/GCWorld/Routing/compare/3.6.2...3.6.2.1)
  - Emergency bug fix for that thing that didn't get quite completely reverted